support reading Argparse instances inside classes
So far, shphinxarg only has support for reading/importing Argparse
instances from global variables/attributes within a module.  However,
there are use cases where Argparse is used inside classes of a module,
not as a global variable.  This is particularly the case for uses
of 'argparse' in the context of CLI / REPL style user interfaces,
such as for example those built using the 'cmd2' module.

This change introduces the ability to specify :func: wit '.' notation
like foo.bar.baz